feat: Look up Metro scope by token
The SourceMapHermes was built specifically for the SourceMap Scope extension as it is used
by the hermes runtime, assuming that we get bytecode offsets (for which the "line" is always 0).
However the scope extension actually comes from the Metro bundler which outputs such scopes also for non-minified JS.

This now exposes a method to get the scope name for a resolved token, no matter the line/column.