Jridgewell index (#54)
The old code improperly found the section to call recurse into. Sourcemap sections behave very similarly to the mappings, and we need to find the section whose offset is `<=` the desired position.

Additionally, the `col` is only subtracted if it is on the offset's line. This matches [Mozilla's behavior](https://github.com/mozilla/source-map/blob/0.6.1/lib/source-map-consumer.js#L1000-L1003).

And as a final bonus, we can binary search it to speed up searches on larger maps.

Co-authored-by: Justin Ridgewell <justin@ridgewell.name>