perf: Add a lazy variant (#1)
In case of most transforms (for ECMAScript), the original input source
map is used only to adjust mappings, and other fields are kept intact.

The _lazy_ variant source maps are source maps used only to call
`adjust_mappings`. It does not deserialize everything. Instead, it only
deserializes the fields required to calculate the new mappings.

This is done in the name of the performance, and the justification for
introducing a new type is at
https://github.com/vercel/next.js/pull/80177#issuecomment-2941464441
<img width="881" alt="image"
src="https://github.com/user-attachments/assets/7ca4b084-ba46-4384-ac25-fefeae0e2513"
/>